Crate process_allocator_test_bootstrap

Functions

main 🔒

A simple bootstrap binary that writes logs out to a file. This is useful for debugging, as normally the ProcessAllocator children logs are piped back to ProcessAllocator. When we are testing what happens when we sigkill ProcessAllocator, we want to see what is happening on the children.